Ticket #55: tab completion vs. spaces and escaping
In the file copy and similar dialogs, filenames are typed as pure strings without
any kind of escaping, and a text entry only takes one filename. However,
splitting at spaces is still performed. Suppose I have a directory named
"/tmp/a bcde". I press F5/F6 on a file, and type "/tmp/a" and then Tab, it
completes to "/tmp/a bcde". This is okay. But if I type "/tmp/a bc" and press
Tab now, it completes to the files in the current directory whose name begins with
"bc", due to the splitting at the space. Similarly, if I already have "/tmp/a bcde/"
in the edit field, it won't complete to subdirectories or files of this directory.
In this case the whole content of the field should be treated as one, not being
split at spaces.

In the command line, splitting at spaces is okay, however, we would need some escaping
and unescaping too, which doesn't happen. (Note: Esc followed by Enter correctly
escapes the filename when it puts it into the cmdline.) Example: type "ls /tmp/a" and
then Tab, you'll have "ls /tmp/a bcdefgh/" in the command line, which clearly won't
list the contents of that directory due to the unescaped space character. Another
example: type "ls /tmp/a\ b" and hit Tab, it won't complete to "ls /tmp/a\ bcde/",
but it should.

Signed-off-by: Slava Zanko <slavazanko@gmail.com>